[CLEANUP] Moved DataFormatHolder to a new package
