+++ src/org/pentaho/agilebi/modeler/Util.gwt.xml
@@ -1,6 +1,5 @@
 <module>
-  <source path="util">
-    <exclude name="**/*Model*"/>
+  <source path="format">
   </source>
 
 </module>
